refactor: CouponType, Coupon, CouponPolicy domain으로 위치 이동
