+++ app/src/main/java/woowacourse/shopping/domain/model/CouponType.kt
@@ -1,8 +1,10 @@
-package woowacourse.shopping.data.model
+package woowacourse.shopping.domain.model
 
 import woowacourse.shopping.domain.repository.CouponPolicy
 
-enum class CouponType(val code: String) {
+enum class CouponType(
+    val code: String,
+) {
     FIXED5000("FIXED5000"),
     FREESHIPPING("FREESHIPPING"),
     MIRACLESALE("MIRACLESALE"),
@@ -11,18 +13,15 @@ enum class CouponType(val code: String) {
     ;
 
     companion object {
-        fun from(code: String): CouponType {
-            return entries.find { it.code == code.trim() } ?: DEFAULT
-        }
+        fun from(code: String): CouponType = entries.find { it.code == code.trim() } ?: DEFAULT
     }
 
-    fun getPolicy(): CouponPolicy {
-        return when (this) {
+    fun getPolicy(): CouponPolicy =
+        when (this) {
             FIXED5000 -> CouponPolicy.Fixed5000
             FREESHIPPING -> CouponPolicy.FreeShipping
             MIRACLESALE -> CouponPolicy.MiracleSale
             BOGO -> CouponPolicy.Bogo
             DEFAULT -> CouponPolicy.Default
         }
-    }
 }
