+++ app/src/main/java/woowacourse/shopping/domain/repository/CouponRepository.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.domain.repository
 
-import woowacourse.shopping.data.model.Coupon
+import woowacourse.shopping.domain.model.Coupon
 import woowacourse.shopping.domain.model.Product
 
 interface CouponRepository {
