+++ app/src/main/java/woowacourse/shopping/presentation/payment/event/CouponEventHandler.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.presentation.payment.event
 
-import woowacourse.shopping.data.model.Coupon
+import woowacourse.shopping.domain.model.Coupon
 
 interface CouponEventHandler {
     fun onCouponCheck(coupon: Coupon)
