+++ app/src/main/java/woowacourse/shopping/presentation/payment/event/CouponEventHandlerImpl.kt
@@ -1,12 +1,11 @@
 package woowacourse.shopping.presentation.payment.event
 
-import woowacourse.shopping.data.model.Coupon
+import woowacourse.shopping.domain.model.Coupon
 import woowacourse.shopping.presentation.payment.PaymentViewModel
 
 class CouponEventHandlerImpl(
     private val viewModel: PaymentViewModel,
-) :
-    CouponEventHandler {
+) : CouponEventHandler {
     override fun onCouponCheck(coupon: Coupon) {
         viewModel.onCheckCoupon(coupon)
     }
