+++ app/src/test/java/woowacourse/shopping/fixture/FakeCouponRepository.kt
@@ -1,10 +1,10 @@
 package woowacourse.shopping.fixture
 
-import woowacourse.shopping.data.model.Coupon
-import woowacourse.shopping.data.model.CouponPolicyContext
 import woowacourse.shopping.data.model.CouponResponse
-import woowacourse.shopping.data.model.CouponType
 import woowacourse.shopping.data.util.TimeProvider
+import woowacourse.shopping.domain.model.Coupon
+import woowacourse.shopping.domain.model.CouponPolicyContext
+import woowacourse.shopping.domain.model.CouponType
 import woowacourse.shopping.domain.repository.CouponRepository
 import woowacourse.shopping.presentation.product.catalog.ProductUiModel
